Moved from shell scripts to gradle tooling to enable easy maven dependency usage (#2)
Rewrote this to use gradle for tooling instead of shell scripting.

My basic goal was to allow for easy local and remote maven deployment with an ordinary maven version so that contracts could be consumed easily as dependencies in gradle projects, and could also easily include dependencies of their own if necessary, to allow contract composition and promote code re-use. This also allowed me to make the code more reliable and fully test it.

I think this meets all the requirements we've put here.

I tried to maintain the idea that few tools are necessary for development. Contract maintainers still don't need to install java or gradle, or know anything about them, and the tools can be run through docker.